services/promotions-display-service/requirements.txt
kafka-python==2.0.2
structlog==23.1.0
requests==2.31.0
pytest==7.4.0
prometheus-client==0.19.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0